+++ HISTORY.md
@@ -1,6 +1,7 @@
 ## 0.0.4
 
-- 
+- Added full image validation for generating loss and dice scores
+- Nested cross-validation added
 
 ## 0.0.3